refactor(mcp): refactor register and discovery (#1763)
