+++ auto-configurations/spring-ai-alibaba-autoconfigure-nacos-mcp-discovery/pom.xml
@@ -29,13 +29,8 @@
 
         <dependency>
             <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-starter-mcp-server</artifactId>
+            <artifactId>spring-ai-starter-mcp-client-webflux</artifactId>
             <optional>true</optional>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-retry</artifactId>
         </dependency>
 
         <dependency>
