+++ auto-configurations/spring-ai-alibaba-autoconfigure-nacos-mcp-discovery/src/main/java/com/alibaba/cloud/ai/autoconfigure/mcp/client/NacosMcpSseClientProperties.java
