+++ auto-configurations/spring-ai-alibaba-autoconfigure-nacos-mcp-register/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -13,9 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-com.alibaba.cloud.ai.autoconfigure.mcp.client.NacosMcpAutoConfiguration
-com.alibaba.cloud.ai.autoconfigure.mcp.client.NacosMcpClientAutoConfiguration
-com.alibaba.cloud.ai.autoconfigure.mcp.client.NacosMcpToolCallbackAutoConfiguration
-com.alibaba.cloud.ai.autoconfigure.mcp.client.NacosMcpTransportBuilderAutoConfiguration
+com.alibaba.cloud.ai.autoconfigure.mcp.register.NacosMcpRegisterAutoConfiguration
