+++ spring-ai-alibaba-bom/pom.xml
@@ -137,13 +137,13 @@
 
             <dependency>
                 <groupId>com.alibaba.cloud.ai</groupId>
-                <artifactId>spring-ai-alibaba-starter-nacos-mcp-client</artifactId>
+                <artifactId>spring-ai-alibaba-starter-nacos-mcp-discovery</artifactId>
                 <version>${project.version}</version>
             </dependency>
 
             <dependency>
                 <groupId>com.alibaba.cloud.ai</groupId>
-                <artifactId>spring-ai-alibaba-starter-nacos-mcp-discovery</artifactId>
+                <artifactId>spring-ai-alibaba-starter-nacos-mcp-register</artifactId>
                 <version>${project.version}</version>
             </dependency>
